introduce a script.tsx to be used in html and index.tsx to serve of entrypoitn to the lib
