+++ index.html
@@ -121,5 +121,5 @@
             </div>
         </div>
     </body>
-    <script type="module" src="/src/index.tsx"></script>
+    <script type="module" src="/src/script.tsx"></script>
 </html>
